Move tomorrowio coordinator to separate module
